Upload a new cache based on EXIT trap
Rather than uploading a new cache as part of the preInstall phase, we now upload
the cache when the builder exits using a bash trap. The rationale for this
change is that the build might fail during the test phase, or compilation of the
very last module. Rather than repeating all the compilation that did succeed, we
still want to resume from that same cache.